Update models v1

+++ backend/src/auth/router.py
@@ -1,13 +1,17 @@
-from fastapi import Depends, Request, APIRouter
+from fastapi import UploadFile, File, APIRouter, Depends, HTTPException, Response
+from typing import List
+import requests
+
+from src.exceptions import CustomException, NotFoundException, NoRightsException
 from src.notifications.utils import send_message
 from src.auth.dependencies import get_current_user
-from fastapi import APIRouter, HTTPException, Response, Depends
+from src.services import MediaRepository
 from src.auth.services import UsersDAO, ArtistDAO, ProducerDAO
 from src.auth.utils import authenticate_user, create_access_token, create_refresh_token, get_hashed_password, verify_password
-from src.auth.schemas import SUser, SArtist, SProducer, SRegisterUser, SLoginUser, SUserBase, SUserUpdate
+from src.auth.schemas import SUser, SArtist, SProducer, SRegisterUser, SLoginUser, SUserBase, SUserUpdate, SUserResponse
 from src.config import settings
-import requests
-from typing import List
+from src.beats.utils import unique_filename
+
 
 auth = APIRouter(
     prefix = "/auth",
@@ -26,29 +30,73 @@ async def get_users() -> List[SUser]:
     response = await UsersDAO.find_all()
     return response
 
-@auth.get('/users/{id}', summary='Get one user by id')
-async def get_user(id: int) -> SUser:
-    response = await UsersDAO.find_one_by_id(id)
-    return response
+@auth.get("/users/{id}", summary="Get a user by ID")
+async def get_one(id: int) -> SUser:
+    user = await UsersDAO.find_one_by_id(id)
+    if not user:
+        raise NotFoundException()
+    
+    user_dict = {
+        "id": user.id,
+        "username": user.username,
+        "email": user.email,
+        "picture_url": user.picture_url,
+        "birthday": user.birthday,
+        "role": user.role,
+        "updated_at": user.updated_at,
+        "created_at": user.created_at
+    }
+
+    return SUser(**user_dict)
+
+@auth.put('/users/picture/{id}', summary='Update image info by id')
+async def update_user_picture(id: int, file: UploadFile = File(...), user: SUser = Depends(get_current_user)):
+    if user.id != id:
+        raise NoRightsException()
+    
+    filename = await unique_filename(file) if file else None
+    picture_url = await MediaRepository.upload_file("PICTURES", filename, file)
+    
+    update_data = {
+        "picture_url": picture_url
+    }
+    await UsersDAO.edit_one(id, update_data)
+    return CustomException(200, f"User {user.username} picture_url has been updated")
 
 @auth.put('/users/{id}', summary='Update user info by id')
-async def update_users(id: int, data: SUserUpdate):
-    update_data = {
-        "username": data.username,
-        "email": data.email,
-        "picture_url": data.picture_url,
-    }
-    response = await UsersDAO.edit_one(id, update_data)
-    return response
+async def update_user(id: int, data: SUserUpdate, user: SUser = Depends(get_current_user)):
+    if not user:
+        raise NotFoundException()
+
+    if user.id != id:
+        raise NoRightsException()
+    
+    update_data = {}
+
+    if data.username:
+        update_data["username"] = data.username
+    if data.email:
+        update_data["email"] = data.email
+    if data.picture_url:
+        update_data["picture_url"] = data.picture_url
+
+    await UsersDAO.edit_one(id, update_data)
+    return update_data
 
 @auth.delete('/users/{id}', summary='Delete user by id')
-async def delete_users(id):
-    response = await UsersDAO.delete(id)
-    return response
+async def delete_users(id: int, user: SUser = Depends(get_current_user)):
+    if not user:
+        raise NotFoundException()
+    
+    if user.id != id:
+        raise NoRightsException()
+    
+    await UsersDAO.delete(id)
+    return CustomException(200, f"User with id {id} has been deleted")
 
 @auth.get('/users/artists', summary='Get all artists')
 async def get_artists() -> List[SArtist]:
-    response = await UsersDAO.find_all_artists()
+    response = await ArtistDAO.find_all()
     return response
 
 @auth.get('/users/artists/{id}', summary='Get one artists by id')
@@ -63,17 +111,17 @@ async def update_artists():
 
 @auth.get('/users/producers', summary='Get all producers')
 async def get_producers():
-    response = await UsersDAO.find_all_producers()
+    response = await ProducerDAO.find_all()
     return response
 
 @auth.get('/users/producers/{id}', summary='Get one producer by id')
 async def get_producer():
-    response = await UsersDAO.find_one_by_id()
+    response = await ProducerDAO.find_one_by_id()
     return response
 
 @auth.put('/users/producers/{id}', summary='Update producer by id')
 async def update_producer():
-    response = await UsersDAO.edit_one()
+    response = await ProducerDAO.edit_one()
     return response
 
 
@@ -97,29 +145,25 @@ async def register(user: SRegisterUser):
     
     print(f"{user_id} FSDEWJKKJLREJKL:")
     
-    # if user.role == "Artist":
-    #     artist_profile_data = {
-    #         "description": "Hi, I'm an artist"
-    #     }
-    #     artist_profile_id = await ArtistDAO.add_one(artist_profile_data)
-    #     print(f"{artist_profile_id} TFGDHJKSADFRWEUIFWEUIMNUIMNEWUIOFEFWEFWWEFUSDFI")
-    #     artist_profile = await ArtistDAO.find_one_by_id(artist_profile_id)
+    if user.role == "Artist":
+        artist_profile_data = {
+            "description": "Hi, I'm an artist"
+        }
+        artist_profile_id = await ArtistDAO.add_one(artist_profile_data)
 
-    #     await UsersDAO.edit_one(user_id, {"artist_profile": artist_profile})
-        
-    # if user.role == "Producer":
-    #     producer_profile_data = {
-    #         "description": "Hi, im a producer",
-    #     }
-    #     producer_profile_id = await ProducerDAO.add_one(producer_profile_data)
-    #     producer_profile = await ProducerDAO.find_one_by_id(producer_profile_id)
+        await UsersDAO.edit_one(user_id, {"artist_profile_id": artist_profile_id})
 
-    #     await UsersDAO.edit_one(user_id, {"producer_profile": producer_profile})
+    if user.role == "Producer":
+        producer_profile_data = {
+            "description": "Hi, I'm a producer"
+        }
+        producer_profile_id = await ProducerDAO.add_one(producer_profile_data)
 
-    # send_message(f"Your account was created, {user.username}", user.email)
+        await UsersDAO.edit_one(user_id, {"producer_profile_id": producer_profile_id})
+
     return user_id
 
-
+     
 @auth.post('/login', summary="Signin")
 async def login(user: SLoginUser, response: Response):
     auth_user = await authenticate_user(email=user.email, password=user.password)
